# be/services/docs/audit-service/swagger.yaml
basePath: /api/v1
info:
  contact: {}
  description: API documentation for audit-service.
  title: Base Golang Audit Service API
  version: "1.0"
paths:
...
  /clients:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List clients
      tags:
      - Clients
    post:
      consumes:
      - application/json
      parameters:
      - description: Create client payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Create client
      tags:
      - Clients
  /clients/{clientId}:
    delete:
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Delete client
      tags:
      - Clients
    get:
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Get client
      tags:
      - Clients
    put:
      consumes:
      - application/json
      parameters:
      - description: Client ID
        in: path
        name: clientId
        required: true
        type: integer
      - description: Update client payload
        in: body
        name: payload
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Update client
      tags:
      - Clients
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"